- fastapi controller OK - removed old file and renamed file structure
