+++ descrivizio-001/ApplicationImage.py
@@ -1,0 +1,4 @@
+from pydantic import BaseModel
+
+class ApplicationImage(BaseModel):
+    content : bytes
