+++ requirements.txt
@@ -9,3 +9,6 @@ datasets~=2.11.0
 requests~=2.28.2
 matplotlib~=3.7.1
 python-decouple~=3.8
+uvicorn~=0.21.1
+fastapi~=0.95.0
+pydantic~=1.10.7
